Simplify transformation saving process
