+++ src/Transformation.tsx
@@ -3,33 +3,13 @@ import React, { ReactElement } from "react";
 import { useState } from "react";
 import "./Transformation.css";
 import CodapFlowErrorDisplay from "./Error";
-import {
-  BaseTransformations,
-  SavedTransformation,
-  SavedTransformationContent,
-  TransformationSaveData,
-} from "./transformation-components/types";
+import { SavedTransformation } from "./transformation-components/types";
 import { PolymorphicComponent } from "./transformation-components/PolymorphicComponent";
-import { createDataInteractive } from "./utils/codapPhone";
 import transformationList, {
+  BaseTransformationName,
   TransformationGroup,
 } from "./transformation-components/transformationList";
 import { useMemo } from "react";
-
-/**
- * Subscribing to this context allows adding new saved transformations
- */
-export const SaveTransformationContext = React.createContext<
-  (
-    name: string,
-    description: string | undefined,
-    d: TransformationSaveData
-  ) => void
->(() => {
-  // If someone tries to subscribe to this context without a valid
-  // provider higher up, then throw an error.
-  throw Error("No Save Transformation Provider Found");
-});
 
 /**
  * Transformation represents an instance of the plugin, which applies a
@@ -44,65 +24,16 @@ function Transformation({
 
   const [errMsg, setErrMsg] = useState<string | null>(null);
 
-  const transformationData: SavedTransformation[] = [
-    { name: "Build Column", content: { base: "Build Column" } },
-    { name: "Compare", content: { base: "Compare" } },
-    { name: "Count", content: { base: "Count" } },
-    { name: "Difference From", content: { base: "Difference From" } },
-    { name: "Filter", content: { base: "Filter" } },
-    { name: "Flatten", content: { base: "Flatten" } },
-    { name: "Running Sum", content: { base: "Running Sum" } },
-    { name: "Running Mean", content: { base: "Running Mean" } },
-    { name: "Running Min", content: { base: "Running Min" } },
-    { name: "Running Max", content: { base: "Running Max" } },
-    { name: "Running Difference", content: { base: "Running Difference" } },
-    { name: "Reduce", content: { base: "Reduce" } },
-    { name: "Group By", content: { base: "Group By" } },
-    { name: "Pivot Longer", content: { base: "Pivot Longer" } },
-    { name: "Pivot Wider", content: { base: "Pivot Wider" } },
-    { name: "Select Attributes", content: { base: "Select Attributes" } },
-    { name: "Sort", content: { base: "Sort" } },
-    { name: "Transform Column", content: { base: "Transform Column" } },
-    { name: "Join", content: { base: "Join" } },
-    { name: "Copy", content: { base: "Copy" } },
-    { name: "Dot Product", content: { base: "Dot Product" } },
-    { name: "Average", content: { base: "Average" } },
-    { name: "Copy Schema", content: { base: "Copy Schema" } },
-    { name: "Combine Cases", content: { base: "Combine Cases" } },
-    { name: "Partition", content: { base: "Partition" } },
-  ];
-
-  function addTransformation(
-    name: string,
-    description: string | undefined,
-    data: TransformationSaveData
-  ) {
-    if (name === "") {
-      return;
-    }
-    // Make sure transformType is non-null
-    if (transformType === null) {
-      return;
-    }
-    // Search for the base value associated with the currently selected transformation
-    const base: BaseTransformations | undefined = transformationData.find(
-      ({ name }) => name === transformType
-    )?.content.base;
-    // Make sure the previous search was successful
-    if (base === undefined) {
-      return;
-    }
-    // Create a new transformation and add it to the list
-    // TODO: can we do this without casting?
-    const content: SavedTransformationContent = {
-      base,
-      data,
-    } as SavedTransformationContent;
-
-    const savedTransformation = { name, description, content };
-    const encoded = encodeURIComponent(JSON.stringify(savedTransformation));
-    createDataInteractive(name, `http://localhost:3000?transform=${encoded}`);
-  }
+  // These are the base transformation types represented as SavedTransformation
+  // objects
+  const baseTransformations: SavedTransformation[] = useMemo(
+    () =>
+      Object.keys(transformationList).map((transform) => ({
+        name: transform,
+        content: { base: transform as BaseTransformationName },
+      })),
+    []
+  );
 
   function typeChange(event: React.ChangeEvent<HTMLSelectElement>) {
     setTransformType(event.target.value);
@@ -163,23 +94,21 @@ function Transformation({
           </select>
         </>
       )}
-      <SaveTransformationContext.Provider value={addTransformation}>
-        {urlTransformation ? (
-          <PolymorphicComponent
-            setErrMsg={setErrMsg}
-            errorDisplay={<CodapFlowErrorDisplay message={errMsg} />}
-            transformation={urlTransformation}
-          />
-        ) : (
-          <PolymorphicComponent
-            setErrMsg={setErrMsg}
-            errorDisplay={<CodapFlowErrorDisplay message={errMsg} />}
-            transformation={transformationData.find(
-              ({ name }) => name === transformType
-            )}
-          />
-        )}
-      </SaveTransformationContext.Provider>
+      {urlTransformation ? (
+        <PolymorphicComponent
+          setErrMsg={setErrMsg}
+          errorDisplay={<CodapFlowErrorDisplay message={errMsg} />}
+          transformation={urlTransformation}
+        />
+      ) : (
+        <PolymorphicComponent
+          setErrMsg={setErrMsg}
+          errorDisplay={<CodapFlowErrorDisplay message={errMsg} />}
+          transformation={baseTransformations.find(
+            ({ name }) => name === transformType
+          )}
+        />
+      )}
     </div>
   );
 }
