+++ src/transformation-components/types.ts
@@ -1,5 +1,6 @@
 import { ReactElement } from "react";
 import { DDTransformationState } from "./DataDrivenTransformation";
+import { BaseTransformationName } from "./transformationList";
 
 /**
  * The content associated with a saved transformation. Includes the base
@@ -7,41 +8,9 @@ import { DDTransformationState } from "./DataDrivenTransformation";
  * formulas, etc.)
  */
 export type SavedTransformationContent = {
-  base:
-    | "Build Column"
-    | "Compare"
-    | "Count"
-    | "Difference From"
-    | "Filter"
-    | "Flatten"
-    | "Running Sum"
-    | "Running Mean"
-    | "Running Min"
-    | "Running Max"
-    | "Running Difference"
-    | "Group By"
-    | "Pivot Longer"
-    | "Pivot Wider"
-    | "Select Attributes"
-    | "Sort"
-    | "Transform Column"
-    | "Copy"
-    | "Copy Schema"
-    | "Join"
-    | "Dot Product"
-    | "Average"
-    | "Combine Cases"
-    | "Reduce"
-    | "Partition";
+  base: BaseTransformationName;
   data?: DDTransformationState;
 };
-
-/**
- *  All valid values of the `base` field of a saved transformation object
- */
-export type BaseTransformations = NonNullable<
-  SavedTransformationContent["base"]
->;
 
 /**
  * All valid save data types
